@@ -1,11 +1,11 @@
 name: receive_sharing_intent
 description: A flutter plugin that enables flutter apps to receive sharing photos, text or url from other apps.
-version: 1.4.2
+version: 1.4.5
 homepage: https://github.com/KasemJaffer/receive_sharing_intent
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
-  flutter: ">=1.10.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.10.0"
 
 dependencies:
   flutter:
