@@ -5,14 +5,14 @@ import 'package:flutter/services.dart';
 
 class ReceiveSharingIntent {
   static const MethodChannel _mChannel =
-      const MethodChannel('receive_sharing_intent/messages');
+  const MethodChannel('receive_sharing_intent/messages');
   static const EventChannel _eChannelMedia =
-      const EventChannel("receive_sharing_intent/events-media");
+  const EventChannel("receive_sharing_intent/events-media");
   static const EventChannel _eChannelLink =
-      const EventChannel("receive_sharing_intent/events-text");
+  const EventChannel("receive_sharing_intent/events-text");
 
-  static Stream<List<SharedMediaFile>> _streamMedia;
-  static Stream<String> _streamLink;
+  static Stream<List<SharedMediaFile>>? _streamMedia;
+  static Stream<String>? _streamLink;
 
   /// Returns a [Future], which completes to one of the following:
   ///
@@ -22,8 +22,8 @@ class ReceiveSharingIntent {
   /// NOTE. The returned media on iOS (iOS ONLY) is already copied to a temp folder.
   /// So, you need to delete the file after you finish using it
   static Future<List<SharedMediaFile>> getInitialMedia() async {
-    final String json = await _mChannel.invokeMethod('getInitialMedia');
-    if (json == null) return null;
+    final json = await _mChannel.invokeMethod('getInitialMedia');
+    if (json == null) return [];
     final encoded = jsonDecode(json);
     return encoded
         .map<SharedMediaFile>((file) => SharedMediaFile.fromJson(file))
@@ -34,7 +34,7 @@ class ReceiveSharingIntent {
   ///
   ///   * the initially stored link (possibly null), on successful invocation;
   ///   * a [PlatformException], if the invocation failed in the platform plugin.
-  static Future<String> getInitialText() async {
+  static Future<String?> getInitialText() async {
     return await _mChannel.invokeMethod('getInitialText');
   }
 
@@ -43,8 +43,8 @@ class ReceiveSharingIntent {
   ///
   /// If the link is not valid as a URI or URI reference,
   /// a [FormatException] is thrown.
-  static Future<Uri> getInitialTextAsUri() async {
-    final String data = await getInitialText();
+  static Future<Uri?> getInitialTextAsUri() async {
+    final data = await getInitialText();
     if (data == null) return null;
     return Uri.parse(data);
   }
@@ -68,12 +68,12 @@ class ReceiveSharingIntent {
   static Stream<List<SharedMediaFile>> getMediaStream() {
     if (_streamMedia == null) {
       final stream =
-          _eChannelMedia.receiveBroadcastStream("media").cast<String>();
+      _eChannelMedia.receiveBroadcastStream("media").cast<String?>();
       _streamMedia = stream.transform<List<SharedMediaFile>>(
-        new StreamTransformer<String, List<SharedMediaFile>>.fromHandlers(
-          handleData: (String data, EventSink<List<SharedMediaFile>> sink) {
+        new StreamTransformer<String?, List<SharedMediaFile>>.fromHandlers(
+          handleData: (String? data, EventSink<List<SharedMediaFile>> sink) {
             if (data == null) {
-              sink.add(null);
+              sink.add([]);
             } else {
               final encoded = jsonDecode(data);
               sink.add(encoded
@@ -85,7 +85,7 @@ class ReceiveSharingIntent {
         ),
       );
     }
-    return _streamMedia;
+    return _streamMedia!;
   }
 
   /// Sets up a broadcast stream for receiving incoming link change events.
@@ -108,7 +108,7 @@ class ReceiveSharingIntent {
     if (_streamLink == null) {
       _streamLink = _eChannelLink.receiveBroadcastStream("text").cast<String>();
     }
-    return _streamLink;
+    return _streamLink!;
   }
 
   /// A convenience transformation of the stream to a `Stream<Uri>`.
@@ -124,11 +124,7 @@ class ReceiveSharingIntent {
     return getTextStream().transform<Uri>(
       new StreamTransformer<String, Uri>.fromHandlers(
         handleData: (String data, EventSink<Uri> sink) {
-          if (data == null) {
-            sink.add(null);
-          } else {
-            sink.add(Uri.parse(data));
-          }
+          sink.add(Uri.parse(data));
         },
       ),
     );
@@ -147,12 +143,12 @@ class SharedMediaFile {
   final String path;
 
   /// Video thumbnail
-  final String thumbnail;
+  final String? thumbnail;
 
   /// Video duration in milliseconds
-  final int duration;
+  final int? duration;
 
-  /// Whether its a video or image
+  /// Whether its a video or image or file
   final SharedMediaType type;
 
   SharedMediaFile(this.path, this.thumbnail, this.duration, this.type);
