@@ -7,6 +7,8 @@ public class SwiftReceiveSharingIntentPlugin: NSObject, FlutterPlugin, FlutterSt
     static let kEventsChannelMedia = "receive_sharing_intent/events-media";
     static let kEventsChannelLink = "receive_sharing_intent/events-text";
     
+    private var customSchemePrefix = "ShareMedia";
+    
     private var initialMedia: [SharedMediaFile]? = nil
     private var latestMedia: [SharedMediaFile]? = nil
     
@@ -16,10 +18,12 @@ public class SwiftReceiveSharingIntentPlugin: NSObject, FlutterPlugin, FlutterSt
     private var eventSinkMedia: FlutterEventSink? = nil;
     private var eventSinkText: FlutterEventSink? = nil;
     
+    // Singleton is required for calling functions directly from AppDelegate
+    // - it is required if the developer is using also another library, which requires to call "application(_:open:options:)"
+    // -> see Example app
+    public static let instance = SwiftReceiveSharingIntentPlugin()
     
     public static func register(with registrar: FlutterPluginRegistrar) {
-        let instance = SwiftReceiveSharingIntentPlugin()
-        
         let channel = FlutterMethodChannel(name: kMessagesChannel, binaryMessenger: registrar.messenger())
         registrar.addMethodCallDelegate(instance, channel: channel)
         
@@ -49,28 +53,69 @@ public class SwiftReceiveSharingIntentPlugin: NSObject, FlutterPlugin, FlutterSt
             result(FlutterMethodNotImplemented);
         }
     }
-    
-    public func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [AnyHashable : Any] = [:]) -> Bool {
-        if let url = launchOptions[UIApplication.LaunchOptionsKey.url] as? URL {
-            return handleUrl(url: url, setInitialData: true)
-        } else if let activityDictionary = launchOptions[UIApplication.LaunchOptionsKey.userActivityDictionary] as? [AnyHashable: Any] { //Universal link
-            for key in activityDictionary.keys {
-                if let userActivity = activityDictionary[key] as? NSUserActivity {
-                    if let url = userActivity.webpageURL {
-                        return handleUrl(url: url, setInitialData: true)
-                    }
-                }
-            }
+
+    // By Adding bundle id to prefix, we'll ensure that the correct application will be openned
+    // - found the issue while developing multiple applications using this library, after "application(_:open:options:)" is called, the first app using this librabry (first app by bundle id alphabetically) is opened
+    public func hasMatchingSchemePrefix(url: URL?) -> Bool {
+        if let url = url, let appDomain = Bundle.main.bundleIdentifier {
+            return url.absoluteString.hasPrefix("\(self.customSchemePrefix)-\(appDomain)")
         }
         return false
     }
     
-    public func application(_ application: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
-        return handleUrl(url: url, setInitialData: false)
+    // This is the function called on app startup with a shared link if the app had been closed already.
+    // It is called as the launch process is finishing and the app is almost ready to run.
+    // If the URL includes the module's ShareMedia prefix, then we process the URL and return true if we know how to handle that kind of URL or false if the app is not able to.
+    // If the URL does not include the module's prefix, we must return true since while our module cannot handle the link, other modules might be and returning false can prevent
+    // them from getting the chance to.
+    // Reference: https://developer.apple.com/documentation/uikit/uiapplicationdelegate/1622921-application
+    public func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [AnyHashable : Any] = [:]) -> Bool {
+        if let url = launchOptions[UIApplication.LaunchOptionsKey.url] as? URL {
+            if (hasMatchingSchemePrefix(url: url)) {
+                return handleUrl(url: url, setInitialData: true)
+            }
+            return true
+        } else if let activityDictionary = launchOptions[UIApplication.LaunchOptionsKey.userActivityDictionary] as? [AnyHashable: Any] {
+            // Handle multiple URLs shared in
+            for key in activityDictionary.keys {
+                if let userActivity = activityDictionary[key] as? NSUserActivity {
+                    if let url = userActivity.webpageURL {
+                        if (hasMatchingSchemePrefix(url: url)) {
+                            return handleUrl(url: url, setInitialData: true)
+                        }
+                        return true
+                    }
+                }
+            }
+        }
+        return true
     }
     
+    // This is the function called on resuming the app from a shared link.
+    // It handles requests to open a resource by a specified URL. Returning true means that it was handled successfully, false means the attempt to open the resource failed.
+    // If the URL includes the module's ShareMedia prefix, then we process the URL and return true if we know how to handle that kind of URL or false if we are not able to.
+    // If the URL does not include the module's prefix, then we return false to indicate our module's attempt to open the resource failed and others should be allowed to.
+    // Reference: https://developer.apple.com/documentation/uikit/uiapplicationdelegate/1623112-application
+    public func application(_ application: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
+        if (hasMatchingSchemePrefix(url: url)) {
+            return handleUrl(url: url, setInitialData: false)
+        }
+        return false
+    }
+    
+    // This function is called by other modules like Firebase DeepLinks.
+    // It tells the delegate that data for continuing an activity is available. Returning true means that our module handled the activity and that others do not have to. Returning false tells
+    // iOS that our app did not handle the activity.
+    // If the URL includes the module's ShareMedia prefix, then we process the URL and return true if we know how to handle that kind of URL or false if we are not able to.
+    // If the URL does not include the module's prefix, then we must return false to indicate that this module did not handle the prefix and that other modules should try to.
+    // Reference: https://developer.apple.com/documentation/uikit/uiapplicationdelegate/1623072-application
     public func application(_ application: UIApplication, continue userActivity: NSUserActivity, restorationHandler: @escaping ([Any]) -> Void) -> Bool {
-        return handleUrl(url: userActivity.webpageURL, setInitialData: true)
+        if let url = userActivity.webpageURL {
+            if (hasMatchingSchemePrefix(url: url)) {
+                return handleUrl(url: url, setInitialData: true)
+            }
+        }
+        return false
     }
     
     private func handleUrl(url: URL?, setInitialData: Bool) -> Bool {
